Documentation and cleanup (#137)
* Replace !!cond with (cond) ? 1 : 0

* Pole documentation

* Some documentation

* Minor cleanup

* Fix some compiler errors

* Move PACKED macro out of global.h

* Minor fixes

* Formatter

* #ifdef out explicit register vars

* Format, and merge input_buffer_1.c with input_buffer.c

* game_7.c: Fix warnings

* Move #include

* Fix some warnings

* Some warnings and cleanup

* Formatter

* Revert bool to !!

* Minor adjustments to sub_8004010

* Formatter

* Remove unused vars

* Warnings, cleanup and documentation

* Fix some warnings

* Some more warnings

* Fix more warnings and add INCREMENT_SCORE_B

* Remove unused vars and other cleanup

* More warnings and cleanup

* Bulk-replace Q_24_8(_TO_INT) and remove a few vars

* Remove unused vars

* Move unused vars

* Fix -Wchar-subscripts warning

* More compiler warnings and cleanup

* Some cam documentation

+++ src/game/stage/debug_text_printer.c
@@ -116,7 +116,7 @@ void Debug_PrintTextAt(char *text, s16 x, s16 y)
 
     for (i = 0; text[i] != '\0'; i++) {
         if (text[i] > ' ') {
-            Sprite *digit = &baseChar[text[i]];
+            Sprite *digit = &baseChar[(u8)text[i]];
 
             digit->x = x;
             digit->y = y;
